Restrict callbacks endpoints to Amazon ip ranges
The callbacks endpoint should only ever be called by Amazon SNS. Amazon
publishes their IP ranges; they're obtainable from an endpoint.
Unfortunately they don't _yet_ specify exactly which SNS uses like they
do for some other services, but they say they will be improving this in
future. We're adding all AMAZON ip ranges for eu-west-1 to ensure we
cover SNS.

This are added by a python script which templates a simple nginx
template. This template is based on the one provided by our docker base
image. The script actually overwrites it. Once templated, the script
reloads nginx to ensure the change is taken into effect.

The script is run every 5 minutes by Cron. There are two cron jobs
added, one that runs periodically and one that runs on reboot/when cron
is started. This ensures we have the most up to date ip ranges
immediately. AWS state that they may change multiple times per week.

Cron isn't started by default, so we start it with syupervisord in the
same way we manage other processes. Supervisord requires that
processes are not demonised, hence why cron is started in the
foreground.

The script logs to file, to allow for any issues to be caught. The file
has been added to the AWS logs agent, and shipped to cloudwatch. This
will allow for an alert or some sort of monitoring to be set up, so we
know if the ip ranges are not being set/updated for any reason.

As we're logging to file, we need to rotate the logfile. As the script
itself isn't being managed by supervisord, we can't use its inbuilt
log rotation. Luckily we have logrotate. So there's a small conf file for
that to ensure that the logfiles from the script never get out of hand.