status checks: add new additional_checks_internal which simply sends an empty signal to check process responsiveness
this is crude, but better than nothing. should be able to detect processes
that have e.g. deadlocked to the point of not being able to handle a signal.
new dependency on psutil to allow us to discover the pid of the process
listening on DM_CLAMD_UNIX_SOCKET_PATH.