Allow overriding development server port
There are times when we might want to change the port that the antivirus
api server listens on in development mode; this commit makes this easier
by adding an additional configuration variable `DM_ANTIVIRUS_API_PORT`
that can be set in the environment.

This commit adds `DM_ANTIVIRUS_API_PORT` to the `.flaskenv`
configuration file; this works because `python-dotenv` does not override
existing environment variables.

We then use `DM_ANTIVIRUS_API_PORT` as the port for the development
server by setting `FLASK_RUN_PORT`; see [the Flask cli docs][1] for more
details on how the development server is configured.

[1]: http://flask.palletsprojects.com/en/1.0.x/cli/