Switch from LXML to defusedxml
LXML is vulnerable to denial of service attacks when parsing
untrusted documents.

Unfortunately, defusedxml lacks a couple of key features which we
were using from LXML; namely good XPath support and the ability
to get an element's namespace.

The only way of identifying an element's namespace in defusedxml
is through the string representation of the element where it will
appear as `"<Element '{http://an.example.namespace.com}ConfirmSubscriptionResponse' at 0x10e7fe4a8>"`

If there is a namespace, extract it from the string representation
and insert it into the search string.

We can no longer get the contents of an element in the XPath, so
we have to use a different XPath in `Element.find()`. This returns
`None` if there's not a match, so wrap each search in a
`try/except` block so we can maintain the behaviour of returning
an empty string if an element isn't found.

https://trello.com/c/HtkcGEV4/811-2-address-ithc-issue-631-xml-parsing-timebox-2-days