Upgrade Flask to 1.1.4
Get the updated version of `digitalmarketplace-utils` to support this.

Pin Flask requirements to 1.1 or greater, but less than
2. Flask 1.1 is the last planned v1 release before 2.0.0, so take
patch versions of 1.1 for security but not 2.0 which we're not ready for
yet.

Flask now controls the version of `itsdangerous` so we can remove that
pinned dependency as well.

Also get rid of pyup comments - we're using dependabot instead now